packages/agent/pyproject.toml
[project]
name = "dessin-coaching-agent"
version = "0.1.0"
description = "鉛筆デッサンコーチングエージェント"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.23.0",
    "google-cloud-aiplatform[agent_engines,adk]>=1.135.0",
    "google-genai>=1.61.0",
    "google-cloud-firestore>=2.23.0",
    "google-cloud-storage>=3.8.0",
    "google-cloud-secret-manager>=2.21.0",
    "fastapi>=0.115.0,<0.124.0",
    "uvicorn[standard]>=0.39.0",
    "httpx>=0.28.0",
    "aiohttp>=3.10.0",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.7.0",
    "structlog>=25.5.0",
    "tenacity>=9.0.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.18",
    "pyjwt>=2.8.0",
    "firebase-admin>=6.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "mypy>=1.14.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line-length (handled by formatter)
    "B008",   # function-call-in-default-argument (FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_any_explicit = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
